EBImage {EBImage}	R Documentation

Package overview

Description

EBImage is an image processing and analysis package for R. Its primary goal is to enable automated analysis of large sets of images such as those obtained in high throughput automated microscopy.

EBImage relies on the Image object to store and process images but also works on multi-dimensional arrays.

Package content

Image methods

Image

as.Image, is.Image, as.raster

colorMode, imageData

getFrame, numberOfFrames

Image I/O, display

readImage, writeImage

display

image

Spatial transforms

resize, flip, flop, transpose

rotate, translate, affine

Image segmentation, objects manipulation

thresh, bwlabel, otsu

watershed, propagate

ocontour

paintObjects, rmObjects, reenumerate

Image enhancement, filtering

normalize

filter2, gblur, medianFilter

Morphological operations

makeBrush

erode, dilate, opening, closing

whiteTopHat, blackTopHat, selfComplementaryTopHat

distmap

floodFill, fillHull

Color space manipulation

rgbImage, channel, toRGB

Image stacking, combining, tiling

stackObjects

combine

tile, untile

Drawing on images

drawCircle

Features extraction

computeFeatures

computeFeatures.basic, computeFeatures.moment, computeFeatures.shape, computeFeatures.haralick

standardExpandRef

Defunct

blur, equalize

drawtext, drawfont

getFeatures, hullFeatures, zernikeMoments

edgeProfile, edgeFeatures,

haralickFeatures, haralickMatrix

moments, cmoments, smoments, rmoments

Authors

Oleg Sklyar, osklyar@ebi.ac.uk, Copyright 2005-2007

Gregoire Pau, gpau@ebi.ac.uk

Wolfgang Huber, huber@ebi.ac.uk

Andrzej Oles, andrzej.oles@embl.de

Mike Smith, msmith@ebi.ac.uk

  European Bioinformatics Institute
  European Molecular Biology Laboratory
  Wellcome Trust Genome Campus
  Hinxton 
  Cambridge CB10 1SD
  UK
  

The code of propagate is based on the CellProfiler with permission granted to distribute this particular part under LGPL, the corresponding copyright (Jones, Carpenter) applies.

The source code is released under LGPL (see the LICENSE file in the package root for the complete license wording).

This library is free software; you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation; either version 2.1 of the License, or (at your option) any later version. This library is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.

See the GNU Lesser General Public License for more details. For LGPL license wording see http://www.gnu.org/licenses/lgpl.html

Examples

  example(readImage)
  example(display)
  example(rotate)
  example(propagate)

[Package EBImage version 4.20.1 Index]