EBImage-defunct {EBImage}	R Documentation

Defunct functions in package ‘EBImage’

Description

These functions are defunct and no longer available.

Details

The following functions are defunct and no longer available; use the replacement indicated below.

animate: display

blur: gblur

drawtext: see package vignette for documentation on how to add text labels to images

drawfont: see package vignette for documentation on how to add text labels to images

getFeatures: computeFeatures

getNumberOfFrames: numberOfFrames

hullFeatures: computeFeatures.shape

zernikeMoments: computeFeatures

edgeProfile: computeFeatures

edgeFeatures: computeFeatures.shape

haralickFeatures: computeFeatures

haralickMatrix: computeFeatures

moments: computeFeatures.moment

cmoments: computeFeatures.moment

rmoments: computeFeatures.moment

smoments: computeFeatures.moment

dilateGreyScale: dilate

erodeGreyScale: erode

openingGreyScale: opening

closingGreyScale: closing

whiteTopHatGreyScale: whiteTopHat

blackTopHatGreyScale: blackTopHat

selfcomplementaryTopHatGreyScale: selfComplementaryTopHat

[Package EBImage version 4.20.1 Index]