abind {EBImage}	R Documentation

Combine Image Arrays

Description

Methods for function abind from package abind useful for combining Image arrays.

Value

An Image object or an array, containing the combined data arrays of the input objects.

Usage

abind(...)

Arguments

...

Arguments to abind

Methods

signature(... = "Image")

This method is defined primarily for the sake of preserving the class of the combined Image objects. Unlike the original abind function, if dimnames for all combined objects are NULL it does not introduce a list of empty dimnames for each dimension.

signature(... = "ANY")

Dispatches to the original abind function.

Author(s)

Andrzej Oleś, andrzej.oles@embl.de, 2017

See Also

combine provides a more convenient interface to merging images into an image sequence. Use tile to lay out images next to each other in a regular grid.

Examples

  f = system.file("images", "sample-color.png", package="EBImage")
  x = readImage(f)
  
  ## combine images horizontally
  y = abind(x, x, along=1)
  display(y)
  
  ## stack images one on top of the other
  z = abind(x, x, along=2)
  display(z)

[Package EBImage version 4.20.1 Index]