colormap {EBImage}	R Documentation

Map a Greyscale Image to Color

Description

Maps a greyscale image to color using a color palette.

Usage

colormap(x, palette = heat.colors(256L))

Arguments

x
an Image object of color mode Grayscale, or an array

palette
character vector containing the color palette

Details

The colormap function first linearly maps the pixel intensity values of x to the integer range 1:length(palette). It then uses these values as indices to the provided color palette to create a color version of the original image.

The default palette contains 256 colors, which is the typical number of different shades in a 8bit grayscale image.

Value

An Image object of color mode Color, containing the color-mapped version of x.

Author(s)

Andrzej Oleś, andrzej.oles@embl.de, 2016

Examples

  x = readImage(system.file("images", "sample.png", package="EBImage"))

  ## posterize an image using the topo.colors palette
  y = colormap(x, topo.colors(8))

  display(y, method="raster")

  ## mimic MatLab's 'jet.colors' colormap
  jet.colors = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000"))
  y = colormap(x, jet.colors(256))

  display(y, method="raster")

[Package EBImage version 4.20.1 Index]