combine {EBImage}	R Documentation

Combine images

Description

Merges images to create image sequences.

Usage

combine(x, y, ...)

Arguments

x
An Image object, an array, or a list containing Image objects and arrays.

y
An Image object or an array.

...
Image objects or arrays.

Details

The function combine uses abind to merge multi-dimensional arrays along the dimension depending on the color mode of x. If x is a Grayscale image or an array, image objects are combined along the third dimension, whereas when x is a Color image they are combined along the fourth dimension, leaving room on the third dimension for color channels.

Value

An Image object or an array.

Author(s)

Gregoire Pau, Andrzej Oles, 2013

See Also

The method abind provides a more flexible interface which allows to specify the dimension along which to combine the images.

Examples

  ## combination of color images
  img = readImage(system.file("images", "sample-color.png", package="EBImage"))[257:768,,]
  x = combine(img, flip(img), flop(img))
  display(x, all=TRUE)

  ## Blurred images
  x = resize(img, 128, 128)
  xt = list()
  for (t in seq(0.1, 5, len=9)) xt=c(xt, list(gblur(x, s=t)))
  xt = combine(xt)
  display(xt, title='Blurred images', all=TRUE)

[Package EBImage version 4.20.1 Index]