display-shiny {EBImage}	R Documentation

Shiny Bindings for display

Description

Output and render functions for using the interactive image viewer within Shiny applications and interactive R Markdown documents.

Usage

displayOutput(outputId, width = "100%", height = "500px")

renderDisplay(expr, env = parent.frame(), quoted = FALSE)

Arguments

outputId
output variable to read from

width, height
Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

expr
An expression that generates the image viewer (typicall through a call to display)

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

See Also

display

Examples

# Only run this example in interactive R sessions
if (interactive()) {
  options(device.ask.default = FALSE)

  require("shiny")
  
  ui <- fluidPage(
  
    # Application title
    titlePanel("Image display"),
  
    # Sidebar with a select input for the image
    sidebarLayout(
      sidebarPanel(
        selectInput("image", "Sample image:", list.files(system.file("images", package="EBImage")))
      ),
  
      # Show a plot of the generated distribution
      mainPanel(
        tabsetPanel(
          tabPanel("Static raster", plotOutput("raster")),
          tabPanel("Interactive browser", displayOutput("widget"))
        )
      )
    )
  
  )
  
  server <- function(input, output) {
  
    img <- reactive({
      f = system.file("images", input$image, package="EBImage")
      readImage(f)
    })
  
    output$widget <- renderDisplay({
      display(img())
    })
  
    output$raster <- renderPlot({
      plot(img(), all=TRUE)
    })
  
  }
  
  # Run the application
  shinyApp(ui = ui, server = server)
}

[Package EBImage version 4.20.1 Index]