equalize {EBImage}	R Documentation

Histogram Equalization

Description

Equalize the image histogram to a specified range and number of levels.

Usage

equalize(x, range = c(0, 1), levels = 256)

Arguments

x
an Image object or an array

range
numeric vector of length 2, the output range of the equalized histogram

levels
number of grayscale levels (Grayscale images) or intensity levels of each channel (Color images)

Details

Histogram equalization is an adaptive image contrast adjustment method. It flattens the image histogram by performing linearization of the cumulative distribution function of pixel intensities.

Individual channels of Color images and frames of image stacks are equalized separately.

Value

An Image object or an array, containing the transformed version of x.

Author(s)

Andrzej Oles, andrzej.oles@embl.de, 2014

See Also

clahe

Examples

  x = readImage(system.file('images', 'cells.tif', package='EBImage'))
  hist(x)
  y = equalize(x)
  hist(y)
  display(y, title='Equalized Grayscale Image')

  x = readImage(system.file('images', 'sample-color.png', package='EBImage'))
  hist(x)
  y = equalize(x)
  hist(y)
  display(y, title='Equalized Grayscale Image')

[Package EBImage version 4.20.1 Index]