fillHull {EBImage}	R Documentation

Fill holes in objects

Description

Fill holes in objects.

Usage

fillHull(x)

Arguments

x
An Image object or an array.

Details

fillHull fills holes in the objects defined in x, where objects are sets of pixels with the same unique integer value.

Value

An Image object or an array, containing the transformed version of x.

Author(s)

Gregoire Pau, Oleg Sklyar; 2007

See Also

bwlabel

Examples

  x = readImage(system.file('images', 'nuclei.tif', package='EBImage'))
  display(x)

  y = thresh(x, 10, 10, 0.05)
  display(y, title='Cell nuclei')

  y = fillHull(y)
  display(y, title='Cell nuclei without holes')

[Package EBImage version 4.20.1 Index]