gblur {EBImage}	R Documentation

Low-pass Gaussian filter

Description

Filters an image with a low-pass Gaussian filter.

Usage

gblur(x, sigma, radius = 2 * ceiling(3 * sigma) + 1, ...)

Arguments

x
An Image object or an array.

sigma
A numeric denoting the standard deviation of the Gaussian filter used for blurring.

radius
The radius of the filter in pixels. Default is 2*ceiling(3*sigma)+1).

...
Arguments passed to filter2.

Details

The Gaussian filter is created with the function makeBrush.

Value

An Image object or an array, containing the filtered version of x.

Author(s)

Oleg Sklyar, osklyar@ebi.ac.uk, 2005-2007

See Also

filter2, makeBrush

Examples

  x = readImage(system.file("images", "sample.png", package="EBImage"))
  display(x)

  y = gblur(x, sigma=8)
  display(y, title='gblur(x, sigma=8)')

[Package EBImage version 4.20.1 Index]