ocontour {EBImage}	R Documentation

Oriented contours

Description

Computes the oriented contour of objects.

Usage

ocontour(x)

Arguments

x
An Image object or an array, containing objects. Only integer values are considered. Pixels of value 0 constitute the background. Each object is a set of pixels with the same unique integer value. Objects are assumed connected.

Value

A list of matrices, containing the coordinates of object oriented contours.

Author(s)

Gregoire Pau, gpau@ebi.ac.uk, 2008

Examples

  x = readImage(system.file("images", "shapes.png", package="EBImage"))
  x = x[1:120,50:120]
  display(x)
  oc = ocontour(x)
  plot(oc[[1]], type='l')
  points(oc[[1]], col=2)

[Package EBImage version 4.20.1 Index]