tile {EBImage}	R Documentation

Tiling/untiling images

Description

Given a sequence of frames, tile generates a single image with frames tiled. untile is the inverse function and divides an image into a sequence of images.

Usage

tile(x, nx=10, lwd=1, fg.col="#E4AF2B", bg.col="gray")
untile(x, nim, lwd=1)

Arguments

x
An Image object, an array or a list of these objects.

nx
The number of tiled images in a row.

lwd
The width of the grid lines between tiled images, can be 0.

fg.col
The color of the grid lines.

bg.col
The color of the background for extra tiles.

nim
A numeric vector of 2 elements for the number of images in both directions.

Details

After object segmentation, tile is a useful addition to stackObjects to have an overview of the segmented objects.

Value

An Image object or an array, containing the tiled/untiled version of x.

Author(s)

Oleg Sklyar, osklyar@ebi.ac.uk, 2006-2007

See Also

stackObjects

Examples

  ## make a set of blurred images
  img = readImage(system.file("images", "sample-color.png", package="EBImage"))[257:768,,]
  x = resize(img, 128, 128)
  xt = list()
  for (t in seq(0.1, 5, len=9)) xt=c(xt, list(gblur(x, s=t)))
  xt = combine(xt)
  display(xt, title='Blurred images')

  ## tile
  xt = tile(xt, 3)
  display(xt, title='Tiles')

  ## untile
  xu = untile(img, c(3, 3))
  display(xu, title='Blocks')

[Package EBImage version 4.20.1 Index]